Support milliseconds for time delta in Test.moveTime

+++ runtime/stdlib/test-framework.go
@@ -75,7 +75,7 @@ type Blockchain interface {
 
 	Reset(uint64)
 
-	MoveTime(int64)
+	MoveTime(float64)
 
 	CreateSnapshot(string) error
